Frontend changes

+++ src/main/java/org/pw/rafalj/crm/repository/contract/impl/ContractRepositoryCriteriaImpl.java
@@ -28,6 +28,6 @@ public class ContractRepositoryCriteriaImpl implements ContractRepository {
         Criteria countCriteria = session.getCurrentSession().createCriteria(Contracts.class);
         countCriteria.setProjection(Projections.rowCount());
 
-        return new PageContainer<>(criteria.list(), (Integer) countCriteria.uniqueResult());
+        return new PageContainer<>(criteria.list(), ((Long) countCriteria.uniqueResult()).intValue());
     }
 }
